% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fmt}
\alias{.fmt}
\title{Format numbers for display}
\usage{
.fmt(x, digits = 3)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{Number of significant digits.}
}
\value{
Character vector.
}
\description{
Format numbers for display
}
\keyword{internal}
