% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{.mv_config}
\alias{.mv_config}
\title{Configure plotly modebar}
\usage{
.mv_config(p)
}
\arguments{
\item{p}{A plotly object.}
}
\value{
A configured plotly object.
}
\description{
Applies standard config to a plotly object.
}
\keyword{internal}
