% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_tabs.R
\name{.mv_match_close_div}
\alias{.mv_match_close_div}
\title{Find the position of a \verb{<div>}'s matching \verb{</div>}}
\usage{
.mv_match_close_div(pos, start)
}
\arguments{
\item{pos}{List as returned by \code{\link[=.mv_div_positions]{.mv_div_positions()}}.}

\item{start}{Position of the specific \verb{<div} whose match to find;
must be an element of \code{pos$open}.}
}
\value{
Position of the matching \verb{</div>} start.
}
\description{
Find the position of a \verb{<div>}'s matching \verb{</div>}
}
