% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{.resid_colours}
\alias{.resid_colours}
\title{Positive/negative residual colours}
\usage{
.resid_colours()
}
\value{
Named list with \code{pos} and \code{neg} colours.
}
\description{
Positive/negative residual colours
}
\keyword{internal}
