% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_object.R
\name{.validate_mv}
\alias{.validate_mv}
\title{Validate an mv object}
\usage{
.validate_mv(x)
}
\arguments{
\item{x}{An mv object.}
}
\value{
Invisible TRUE if valid; otherwise stops with error.
}
\description{
Checks that required columns are present in data.frame
components.
}
\keyword{internal}
