% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{mv_bottom_label}
\alias{mv_bottom_label}
\title{Bottom annotation label}
\usage{
mv_bottom_label(label, offset = 0)
}
\arguments{
\item{label}{Text string.}

\item{offset}{Vertical offset from bottom.}
}
\value{
A list for use in \code{layout(annotations = ...)}.
}
\description{
Bottom annotation label
}
