% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{mv_colours}
\alias{mv_colours}
\title{modelVis default colour palette}
\usage{
mv_colours(n = NULL)
}
\arguments{
\item{n}{Number of colours needed. If NULL, returns the
full palette.}
}
\value{
Character vector of hex colour codes.
}
\description{
Returns a vector of colours for general use in plots.
}
