% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mv_ensure_ci}
\alias{mv_ensure_ci}
\title{Add CI columns if missing}
\usage{
mv_ensure_ci(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
Modified data.frame with lwr/upr columns.
}
\description{
Ensures lwr and upr columns exist, setting to NA if
not already present.
}
