% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{mv_fleet_colours}
\alias{mv_fleet_colours}
\title{Fleet colour palette}
\usage{
mv_fleet_colours(fleet_names)
}
\arguments{
\item{fleet_names}{Character vector of fleet names.}
}
\value{
Named character vector of hex colours.
}
\description{
Returns a named vector of colours for fleets.
}
