% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{mv_kobe_colours}
\alias{mv_kobe_colours}
\title{Status zone colours for Kobe plot}
\usage{
mv_kobe_colours()
}
\value{
Named character vector with quadrant colours.
}
\description{
Status zone colours for Kobe plot
}
