% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{mv_layout}
\alias{mv_layout}
\title{Standard plotly layout options}
\usage{
mv_layout()
}
\value{
A list suitable for \code{plotly::layout()}.
}
\description{
Returns a list of layout settings for consistent
appearance across plots.
}
