% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_tabs.R
\name{mv_lazy_tabs}
\alias{mv_lazy_tabs}
\title{Make non-default flexdashboard pages/tabs lazy-render}
\usage{
mv_lazy_tabs(html_path)
}
\arguments{
\item{html_path}{Path to a rendered flexdashboard HTML file.}
}
\value{
Invisibly, \code{html_path}.
}
\description{
Make non-default flexdashboard pages/tabs lazy-render
}
\keyword{internal}
