% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{mv_left_label}
\alias{mv_left_label}
\title{Left annotation label}
\usage{
mv_left_label(label, offset = 0)
}
\arguments{
\item{label}{Text string.}

\item{offset}{Horizontal offset from left edge.}
}
\value{
A list for use in \code{layout(annotations = ...)}.
}
\description{
Left annotation label
}
