% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kobe.R
\name{mv_plot_kobe}
\alias{mv_plot_kobe}
\title{Kobe phase plot}
\usage{
mv_plot_kobe(
  biomass,
  mortality,
  years,
  xlab = "B/Bmsy",
  ylab = "F/Fmsy",
  title = NULL
)
}
\arguments{
\item{biomass}{Numeric vector of B/Bmsy ratios by year.}

\item{mortality}{Numeric vector of F/Fmsy or U/Umsy
ratios by year.}

\item{years}{Integer vector of years.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{title}{Plot title.}
}
\value{
A plotly object.
}
\description{
Creates a plotly Kobe plot showing stock status
relative to reference points, with coloured quadrants.
}
