% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{mv_right_label}
\alias{mv_right_label}
\title{Right annotation label}
\usage{
mv_right_label(label, offset = 0)
}
\arguments{
\item{label}{Text string.}

\item{offset}{Horizontal offset from right edge.}
}
\value{
A list for use in \code{layout(annotations = ...)}.
}
\description{
Right annotation label
}
