% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{mv_sex_colours}
\alias{mv_sex_colours}
\title{Sex colour palette}
\usage{
mv_sex_colours(sex_names)
}
\arguments{
\item{sex_names}{Character vector of sex names.}
}
\value{
Named character vector of hex colours.
}
\description{
Returns a named vector of colours for sexes.
}
