% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_array.R
\name{mv_tidy_3d}
\alias{mv_tidy_3d}
\title{Tidy a 3D array}
\usage{
mv_tidy_3d(
  x,
  dim_names = c("d1", "d2", "d3"),
  dim_labels = list(NULL, NULL, NULL),
  value_col = "est"
)
}
\arguments{
\item{x}{A 3D array.}

\item{dim_names}{Character vector of length 3 giving
column names for each dimension.}

\item{dim_labels}{List of length 3 giving labels for
each dimension. NULL elements use dimnames or seq.}

\item{value_col}{Name for the value column.}
}
\value{
A data.frame in long format.
}
\description{
Converts a 3D array into long-format data.frame.
}
