% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_array.R
\name{mv_tidy_4d}
\alias{mv_tidy_4d}
\title{Tidy a 4D array}
\usage{
mv_tidy_4d(
  x,
  dim_names = c("d1", "d2", "d3", "d4"),
  dim_labels = list(NULL, NULL, NULL, NULL),
  value_col = "est"
)
}
\arguments{
\item{x}{A 4D array.}

\item{dim_names}{Character vector of length 4.}

\item{dim_labels}{List of length 4.}

\item{value_col}{Name for the value column.}
}
\value{
A data.frame in long format.
}
\description{
Tidy a 4D array
}
