% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_array.R
\name{mv_tidy_5d}
\alias{mv_tidy_5d}
\title{Tidy a 5D array}
\usage{
mv_tidy_5d(
  x,
  dim_names = paste0("d", 1:5),
  dim_labels = vector("list", 5L),
  value_col = "est"
)
}
\arguments{
\item{x}{A 5D array.}

\item{dim_names}{Character vector of length 5.}

\item{dim_labels}{List of length 5.}

\item{value_col}{Name for the value column.}
}
\value{
A data.frame in long format.
}
\description{
Tidy a 5D array
}
