% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mv_year_seq}
\alias{mv_year_seq}
\title{Create a year sequence from model dimensions}
\usage{
mv_year_seq(fYear, nT)
}
\arguments{
\item{fYear}{First year (integer).}

\item{nT}{Number of time steps.}
}
\value{
Integer vector of years.
}
\description{
Create a year sequence from model dimensions
}
