% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp.R
\name{mv_plot_comp_avg}
\alias{mv_plot_comp_avg}
\title{Average composition fit}
\usage{
mv_plot_comp_avg(df, bin_col = "age", ylab = "Mean proportion", title = NULL)
}
\arguments{
\item{df}{A data.frame with bin, \code{obs}, \code{pred}
columns.}

\item{bin_col}{Name of the bin column.}

\item{ylab}{Y-axis label.}

\item{title}{Plot title.}
}
\value{
A plotly object.
}
\description{
Creates a plot of mean observed and predicted
composition across all years.
}
